fixes...important comments regarding OffsetScreenPageLimit ahead:
using a new OffsetScreenPageLimit to avoid fragments reset.
cardFragment is not ready yet for fragment reset, minor corrections are needed for this to be possible.
with reset (using the default offset) the user looses track of the card stack when entering some menus but not all. This can be confusing. Not sure if there is an alternative to this, possible by keeping all data outside the fragment but this would probably require heavy refactoring and thus can't be applied yet...

IMPORTANT NOTE:
The fragments should be ready in case the fragment is reset.
currently they are not ready, at least cardFragment is not !!!
code SHOULD be corrected later to support this!!!